requirements.txt
beautifulsoup4==4.10.0
certifi==2023.7.22
charset-normalizer==2.0.12
idna==3.4
oauthlib==3.2.2
python-dotenv==1.0.0
requests==2.26.0
requests-oauthlib==1.3.1
soupsieve==2.5
splitwise==3.0.0
urllib3==1.26.16
